Switched from blessed to curses.
Even after the optimizations, there was tearing visible when the player moved horizontally. curses's use of screen buffers (windows) eliminates this. As a bonus, we are now also outsourcing to curses the task of choosing which characters to send to the terminal in order to redraw the screen as quickly as possible.